Created a BaseImportWorkflowResource class for use in dictionary and media resources
